% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authModule.R
\name{authServer}
\alias{authServer}
\title{MSAL Auth Serverside}
\usage{
authServer(id)
}
\arguments{
\item{id}{id of the Shiny module}
}
\description{
The Serverside that retrieves the information from the accessToken (see auth.js), then returns it as a reactive
}
