% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authUI.R
\name{authUI}
\alias{authUI}
\title{MSAL Auth UI}
\usage{
authUI(id = NULL)
}
\arguments{
\item{id}{id of the Shiny module}
}
\description{
This UI is used to inject the javascript into de the headers of the page
}
