% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_access_token.R
\name{find_jwk_by_kid}
\alias{find_jwk_by_kid}
\title{Find the JSON Web Key matching the key id in the provided JSON Web Token}
\usage{
find_jwk_by_kid(kid, keys)
}
\arguments{
\item{kid}{key id - a string taken form the JWT Token header. Used to pick the right key from the JWKS}

\item{keys}{a list of public keys (from JWKS) from Microsoft}
}
\value{
The matching key or NULL
}
\description{
This function is needed so we can determine which key to use to verify this token signature.
}
