% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_access_token.R
\name{get_jwks}
\alias{get_jwks}
\title{Get the JSON Web Key Set from the discovery document}
\usage{
get_jwks(tenant_id, token_version)
}
\arguments{
\item{tenant_id}{tenant id from the access token.}

\item{token_version}{version of accesstoken (1.0 or 2.0), to be passed along to \code{get_open_config}}
}
\value{
JSON. The JSON contains public keys in asymmetric cryptography (RSA in this case) that Microsoft uses as a private key to sign
JWT when they issue them.
}
\description{
using the openID Connect discovery document, retrieve the keys from the JSON Web Key Set.
This is a standardized JSON format that contains all public keys a provider (here: Microsoft Entra ID) is currently using.
These are retrieved by reading jwks_uri from the discovery document. Here Microsoft publishes all the current public keys used to sign access tokens.
Your backend never sees the private key. Instead, you use the corresponding public key from JWKS to verify the signature of the JSON Web Tokens (JWT).
}
