% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_access_token.R
\name{get_openid_config}
\alias{get_openid_config}
\title{openId config ophalen/discovery document}
\usage{
get_openid_config(tenant_id, token_version)
}
\arguments{
\item{tenant_id}{tenant id from the access token.}

\item{token_version}{Version of token, determines from which url the config needs to be retrieved}
}
\value{
JSON. OpenID Connect discovery document (JSON with settings, claims and endpoints) of Microsoft Entra ID for given tenant_id
}
\description{
retrieve the OpenID-config for a tenant. The returned information of this function is used to validate:
\itemize{
\item Who issued the token (issuer)
\item Where the public keys are (jwks_uri) to check the signature of the token
\item Which endpoints and features are supported by the provider
}

This function is only used internally and will not be exported.
}
