% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_access_token.R
\name{validate_access_token}
\alias{validate_access_token}
\title{Validate Access Token MSAL}
\usage{
validate_access_token(token, tenant_id, app_id, skew_secs = 300)
}
\arguments{
\item{token}{retrieved access token at starting the application which needs to be checked}

\item{tenant_id}{tenant ID given from the msal config. Can be NULL, will then not be checked but replaced with common.}

\item{app_id}{App ID from Azure (equals client ID, can be found in the MSAL config)}

\item{skew_secs}{Allowed clock skew in seconds when checking expiration timestamp and “not before” timestamp in seconds. Default = 300 secs = 5 minutes}
}
\value{
Either a JSON object with userInfo or FALSE if not all the checks have passed
}
\description{
We want to validate the access token given to the input of the application. This function helps in doing so.
The general idea:
Client logs in → gets access token from Entra ID.
Client sends token to your backend.
Backend reads token header → sees kid (which key was used to sign it).
Backend calls get_jwks() → gets current keys.
Backend finds the key with matching kid.
Backend uses that key to verify the signature.
Backend checks claims ((aud), iss, exp, tid, etc.).
If everything matches → token is valid → allow access.
}
